[Refactor] DB 업데이트, 삭제 후 S3 객체 삭제로 로직 변경
