+++ src/domain/mypage/repository/producer/createProducerPortfolioByUserId.ts
@@ -2,7 +2,7 @@ import prisma from '../../../../global/config/prismaClient';
 import { getAudioDurationInSeconds } from 'get-audio-duration';
 import convertCategory from '../../../../global/modules/convertCategory';
 import config from '../../../../global/config';
-import getS3OneBeatObject from '../../../../global/modules/S3Object/get/getOneBeatObject';
+import getS3OneBeatObject from '../../../../global/modules/S3Object/get/getOneBeat';
 import { PortfolioCreateDTO } from '../../interfaces';
