+++ src/global/modules/S3Object/get/getOneBeat.ts
